WIP explanation of test suites
Working on multiple scripts in a project is a bit painful. Found a way
to run scripts that are not in the .cabal file, 'cabal exec runhaskell <scriptname>' but can't add it to
the head of the script.

Adding each step to the .cabal file is possible, but leads to lot of
(unavoidable) duplication.